+++ arcane/ceapart/src/arcane/tests/srcs.cmake
@@ -14,6 +14,8 @@ set(ARCANE_SOURCES
 set(ARCANE_MATERIAL_SOURCES
   HyodaMixedCellsUnitTest.cc
   MeshMaterialTesterModule.cc
+  MeshMaterialTesterModule_Init.cc
+  MeshMaterialTesterModule_Samples.cc
   MeshMaterialTesterModule.h
   MeshMaterialSyncUnitTest.cc
   MeshMaterialSimdUnitTest.cc
